requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
openpyxl==3.1.5
pdfplumber==0.11.4
python-docx==1.1.2
Pillow==10.4.0
pytesseract==0.3.13
beautifulsoup4==4.12.3
pandas==2.2.2
gigachat==0.1.35
aiofiles==24.1.0
